refactor: Improve state management and type safety
Removes unused state variables `audioUrl` and `panelTimings` from `App.tsx`.
Adds `@types/node` to dev dependencies for better TypeScript support.
Enhances error handling and validation for API responses in `geminiService.ts` and `animationService.ts`.
Ensures correct type casting for video Blob creation in `videoService.ts`.
Cleans up unused API key in `.env.local`.

+++ src/services/videoService.ts
@@ -196,5 +196,5 @@ export async function renderVideo(
         console.error("Could not clean up all files, but proceeding.", e);
     }
     
-    return new Blob([data], { type: 'video/mp4' });
+    return new Blob([data as Uint8Array], { type: 'video/mp4' });
 }
